storage/txn: remove Executor
